Remove broken/misleading language build from AppVeyor
Currently, if you try to make AppVeyor use a different Python version, Pipenv
sees the `python_version` requirement in the `Pipfile` and just runs everything
with Python 3.7 anyway.

Until we get set up with Tox or something there doesn't seem to be a simple way
to fix that. So instead, just remove the other builds so as not to be
misleading about which versions are tested